requirements.txt
autopep8==1.7.0
black==22.10.0
click==8.1.3
joblib==1.2.0
mypy-extensions==0.4.3
numpy==1.23.3
pandas==1.5.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
